Replace WebpackCleanupPlugin by CleanObsoleteChunks
WebpackCleanupPlugin has a issue that it tries to delete everything that
does not belong to the current compile output. However, sometimes
webpack just doesn't output anything because it assumes that some assets
are already compiled on previous compilation.

CleanObsoleteChunks along with an option to clean the whole output on
start `watch` command is the best combination so far, while deleting old
files at runtime and ensuring the output is always clean when we start
watching.