Make global JS file load by JS entry point
This is a major change since last rewrite. We will have only one
javascript entrypoint. This can turn the the entrypoint file much more
powerful, because it controls the whole script execution of the page.

The only downsides I can see now are:

* Additional include that every action (entrypoint) must have pointing
to the global script, otherwise no global stuff (layout, analytics) will
load;
* Difficulty to share cache across multiple pages on the client-side.
This can be easily reverted by making global.js an entrypoint and adding
a reference to load it on `_head.html.erb`, but let's see how it goes.